handle nullable for csharp property
